Add emotion feature and sync audio with text
